requirements.txt
faiss-cpu
transformers
torch
sentence-transformers
huggingface-hub
streamlit
PyPDF2
dotenv